ANgulars happening
Okay, so the reason for no angulars was that I was missing values for
their equations. Now you can see small differences... but they're still
way too small to be realistic.

Maybe it's the logification? I might need to go with Mats' way, scaling
down appropriately by largest force. It shouldn't be that hard, either.